moved util.js from folder api -> utils, updated imports
