+++ backend/service/emailService.js
@@ -13,6 +13,7 @@ const sendEmail = async (to, subject, html) => {
       subject,
       html,
     });
+    console.log("Email sent", subject);
     return response;
   } catch (error) {
     console.error('Error sending email:', error);
