+++ frontend/src/lib/api/commentsApi.js
@@ -1,4 +1,4 @@
-import { makeOption } from "./util.js";
+import { makeOption } from "../utils/util.js";
 
 const BASE_URL = import.meta.env.VITE_API_URL ? `${import.meta.env.VITE_API_URL}/comments` : '/api/comments';
